# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Escrituras",
        "description": "API para acceder a las escrituras de La Iglesia de Jesucristo de los Santos de los Últimos Días",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/pericopas/{id}": {
            "get": {
                "tags": [
                    "Pericopas"
                ],
                "summary": "Obtiene una perícopa específica",
                "operationId": "2c3e76587aaa4b2dab11f726a90dab31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la perícopa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perícopa encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pericopa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perícopa no encontrada"
                    }
                }
            }
        },
        "/api/pericopas/{id}/versiculos": {
            "get": {
                "tags": [
                    "Pericopas"
                ],
                "summary": "Lista todos los versículos de una perícopa",
                "operationId": "a97eb13e2e67f0b0f58f15e12c8c6163",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la perícopa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de versículos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Versiculo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perícopa no encontrada"
                    }
                }
            }
        },
        "/api/pericopas/{id}/capitulo": {
            "get": {
                "tags": [
                    "Pericopas"
                ],
                "summary": "Obtiene el capítulo al que pertenece la perícopa",
                "operationId": "95d6b773c354e98096be170f2d310844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la perícopa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capítulo obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Capitulo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Perícopa no encontrada"
                    }
                }
            }
        },
...
        "/api/versiculos/{id}": {
            "get": {
                "tags": [
                    "Versiculos"
                ],
                "summary": "Obtiene un versículo específico",
                "operationId": "95c0142ad6c7669900d64689fe9a4445",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del versículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Versículo encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Versiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Versículo no encontrado"
                    }
                }
            }
        },
        "/api/versiculos/{id}/pericopa": {
            "get": {
                "tags": [
                    "Versiculos"
                ],
                "summary": "Obtiene la perícopa a la que pertenece el versículo",
                "operationId": "a238b101499260a82f11f386523de334",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del versículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perícopa obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pericopa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Versículo no encontrado"
                    }
                }
            }
        },
        "/api/versiculos/{id}/capitulo": {
            "get": {
                "tags": [
                    "Versiculos"
                ],
                "summary": "Obtiene el capítulo al que pertenece el versículo",
                "operationId": "c5229fc932dcc035d208de0abfd02c73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del versículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capítulo obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Capitulo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Versículo no encontrado"
                    }
                }
            }
        },
...
        "/api/volumenes/{id}": {
            "get": {
                "tags": [
                    "Volumenes"
                ],
                "summary": "Obtiene un volumen específico",
                "operationId": "09bab61eec73e7ea389a6c97bf0de1f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del volumen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Volumen encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Volumen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Volumen no encontrado"
                    }
                }
            }
        },
        "/api/volumenes/{id}/divisiones": {
            "get": {
                "tags": [
                    "Volumenes"
                ],
                "summary": "Lista todas las divisiones de un volumen",
                "operationId": "6532f686f7c49776c120c942e1cea896",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del volumen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de divisiones obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Division"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Volumen no encontrado"
                    }
                }
            }
        },
        "/api/volumenes/{id}/libros": {
            "get": {
                "tags": [
                    "Volumenes"
                ],
                "summary": "Lista todos los libros de un volumen",
                "operationId": "10ef9b4b1d05c08dccc054a87511cfa5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del volumen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de libros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Libro"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Volumen no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Volumen": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Division": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "volumen_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Libro": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "division_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Capitulo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "num_capitulo": {
                        "type": "integer"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "libro_id": {
                        "type": "integer"
                    },
                    "parte_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Pericopa": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "titulo": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "capitulo_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Versiculo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "contenido": {
                        "type": "string"
                    },
                    "num_versiculo": {
                        "type": "integer"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "capitulo_id": {
                        "type": "integer"
                    },
                    "pericopa_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}